final commit

+++ CalculatorProblemV2/src/main/java/com/usama/calculatorproblemv2/entity/OutputParameters.java
@@ -5,32 +5,32 @@ import jakarta.persistence.Embeddable;
 @Embeddable
 public class OutputParameters {
 
-	private String parameterName;
-	private Integer parameterValue;
+    private String parameterName;
+    private Double parameterValue;
 
-	public OutputParameters() {
-	}
+    public OutputParameters() {
+    }
 
-	public OutputParameters(String parameterName, Integer parameterValue) {
-		super();
-		this.parameterName = parameterName;
-		this.parameterValue = parameterValue;
-	}
+    public OutputParameters(String parameterName, Double parameterValue) {
+	super();
+	this.parameterName = parameterName;
+	this.parameterValue = parameterValue;
+    }
 
-	public String getParameterName() {
-		return parameterName;
-	}
+    public String getParameterName() {
+	return parameterName;
+    }
 
-	public void setParameterName(String parameterName) {
-		this.parameterName = parameterName;
-	}
+    public void setParameterName(String parameterName) {
+	this.parameterName = parameterName;
+    }
 
-	public Integer getParameterValue() {
-		return parameterValue;
-	}
+    public Double getParameterValue() {
+	return parameterValue;
+    }
 
-	public void setParameterValue(Integer parameterValue) {
-		this.parameterValue = parameterValue;
-	}
+    public void setParameterValue(Double parameterValue) {
+	this.parameterValue = parameterValue;
+    }
 
 }
